slight correction

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@
  */
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t o, w, len;
+	int o, w, len;
 
 
 	if (!filename)
